+++ src/index.ts
@@ -17,6 +17,7 @@ export { default as Toast } from './Toast'
 export { default as Tooltip } from './Tooltip'
 export { default as Typography } from './Typography'
 export { default as IconFactory } from './IconFactory'
+export { default as Dropdown } from './Dropdown'
 
 // Types
 export type { IconName } from './IconFactory'