+++ CHANGELOG.md
@@ -15,10 +15,10 @@ Types of changes are:
 ## Unreleased
 
 ### Changed
-- Change label option type from string to ReactNode
+- Change `Dropdown component` label option type from string to ReactNode
 
 ### Added
-- Add WithCustomLabel story with custom option labels
+- Add WithCustomLabel story with custom option labels to `Dropdown component` stories
 
 
 ## 2.0.5 - 2025-08-25